refactor(api-bundle): shim ManageApiToken token class as deprecated supertype
Replace the class_alias shim with a real deprecated ManageApiToken base class
that KubernetesServiceAccountToken extends. This keeps the old FQN a valid
type-hint for the produced token, gives IDE/static-analysis @deprecated
discoverability, and works under classmap-authoritative autoloading (which a
class_alias-only file does not).
